requirements.txt
aiogram==3.10.0
jpype1==1.5.0
requests==2.32.3
lxml==5.2.2
beautifulsoup4==4.12.3
tabula-py==2.9.3